+++ src/mock/user-rank.js
@@ -1,3 +1,3 @@
-import {getRandomIntegerNumber} from '../utils.js';
+import {getRandomIntegerNumber} from '../utils/common.js';
 
 export const getUserRank = () => getRandomIntegerNumber(0, 100);
