+++ src/utils/common.js
@@ -1,0 +1,1 @@
+export const getRandomIntegerNumber = (min, max) => min + Math.round(max * Math.random());
