+++ src/utils/const.js
@@ -1,0 +1,11 @@
+export const FilmCount = {
+  ALL: 15,
+  LIST: 5,
+  EXTRA: 2,
+};
+
+export const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`,
+  AFTEREND: `afterend`,
+};
